# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Agent AI System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/gateway/agents/discover": {
      "post": {
        "tags": [
          "gateway"
        ],
        "summary": "Discover cloud agents",
        "operationId": "gateway_discover_api_v1_gateway_agents_discover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayDiscoverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayDiscoveryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gateway/agents/{agent_id}/message/send": {
      "post": {
        "tags": [
          "gateway"
        ],
        "summary": "Send a synchronous message to an agent",
        "operationId": "gateway_send_api_v1_gateway_agents__agent_id__message_send_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewaySendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gateway/agents/{agent_id}/message/stream": {
      "post": {
        "tags": [
          "gateway"
        ],
        "summary": "Stream a message to an agent (SSE)",
        "operationId": "gateway_stream_api_v1_gateway_agents__agent_id__message_stream_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewaySendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gateway/agents/{agent_id}/card": {
      "get": {
        "tags": [
          "gateway"
        ],
        "summary": "Get an agent's card with masked URL",
        "operationId": "gateway_get_card_api_v1_gateway_agents__agent_id__card_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GatewayCardResponse": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "agent_card": {
            "additionalProperties": true,
            "type": "object",
            "title": "Agent Card"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "agent_card"
        ],
        "title": "GatewayCardResponse"
      },
      "GatewayDiscoverRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100,
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "GatewayDiscoverRequest"
      },
      "GatewayDiscoveryAgentResult": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "agent_card": {
            "additionalProperties": true,
            "type": "object",
            "title": "Agent Card"
          },
          "match_score": {
            "type": "number",
            "title": "Match Score"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "agent_card",
          "match_score"
        ],
        "title": "GatewayDiscoveryAgentResult",
        "description": "Discovery result enriched with agent_id for gateway consumers."
      },
      "GatewayDiscoveryResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "agents": {
            "items": {
              "$ref": "#/components/schemas/GatewayDiscoveryAgentResult"
            },
            "type": "array",
            "title": "Agents"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "query",
          "agents",
          "count"
        ],
        "title": "GatewayDiscoveryResponse",
        "description": "Gateway discovery response with agent_id on each result."
      },
      "GatewaySendRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "message": {
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "GatewaySendRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}